Parser: Increase parsing efficiency; add attributes
This commit introduces two new attributes, game_id and line_number, to
generated Event objects. Between the game_id, a unique hash of all the
game's lines, and the line_number, the position of each Event in the
game log, all events can be uniquely identified. This change is
necessary for database storage and retrieval to be possible.