Add special GameEnd event at end of log; improve parser behavior
A special GameEnd event will now fire at the end of a parsed log,
regardless of the log's completeness. This specialized event will
contain statistics from the entire log, such as how long a Player has
been connected, how long a Player has played as a particular role, and
so on.

Due to the GameEnd event bookending the game events with MapLoad at the
start, games can be more efficiently distinguished from one another,
and round divisions have been changed to reflect this.

Additionally, the parser timestamp regex has been migrated to a
constant and changed from a DateTime to a Time in preparation for a
move away from Sequel.