Assume player#in_game? and Player#connected based on Events
If a Player has performed a particular action (i.e. a PlayerSay
indicating they have spoken in the game chat) it can be assumed that
they have connected and/or entered the game. To ensure the accuracy of
Player#durations, assume these events have occurred if they have not
been explicitly triggered by a PlayerConnect or PlayerEnterGame,
respectively.